src/test/java/_0851_0900/_873_Length_of_Longest_Fibonacci_Subsequence/SolutionITest.java
package _0851_0900._873_Length_of_Longest_Fibonacci_Subsequence;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 9/13/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int[] input = {1, 2, 3, 4, 5, 6, 7, 8};
    assertEquals(5, solution.lenLongestFibSubseq(input));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final int[] input = {1, 3, 7, 11, 12, 14, 18};
    assertEquals(3, solution.lenLongestFibSubseq(input));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase3(Solution solution) {
    final int[] input = {1, 3, 5};
    assertEquals(0, solution.lenLongestFibSubseq(input));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}